+++ js/create-objects.js
@@ -1,0 +1,6 @@
+import {
+  CreateArraySimilarAdsNearby,
+  COUNT_OF_AD
+} from './data.js';
+
+CreateArraySimilarAdsNearby(COUNT_OF_AD);
